+++ packages/router/types/side-effects/title.d.ts
@@ -1,3 +1,3 @@
-import { SideEffect, Emitted } from "@curi/types";
+import { Observer, Emitted } from "@curi/types";
 export declare type TitleBuilder = (emitted: Emitted) => string;
-export default function title(callback: TitleBuilder): SideEffect;
+export default function title(callback: TitleBuilder): Observer;