+++ website/src/pages/Guides/sync-or-async.js
@@ -68,8 +68,8 @@ function SyncAndAsyncGuide() {
           <li>
             <p>
               If the initial route that matches is async and you try to render
-              immediately, the <IJS>response</IJS> will be <IJS>null</IJS>. You
-              can wait to render until the initial response is ready with{" "}
+              immediately, the <IJS>response</IJS> will be <IJS>undefined</IJS>.
+              You can wait to render until the initial response is ready with{" "}
               <IJS>router.once</IJS>.
             </p>