+++ packages/router/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Next
+
+* `response` and `navigation` are `undefined` until initial response is finalized.
+
 ## 2.0.0-beta.20
 
 * Merge the `aria-live` (now `announce`), `scroll`, and `title` side effects into package.